New config file
There is a new config file 'taginfo-config.json'. Copy it one directory level
above where it is in the distribution and change it to your needs. There are
multiple changes in the code to get settings from the config file. This
should make running your own instance a lot easier.