Fix language switching.
Switching between languages didn't work in some constellations. This seems to
be because of caching of the pages in the browser. I think the browser should
not use the cached page, because it has a different cookie, but it did. So
I changed the expire on normal web pages to 0. (API pages still have a longer
expire.)