Empty table issue cleanup.
Sometimes there are not results for some query and you'll end up with an empty
table. It would be nice if no empty table appears in this case but instead some
nice message. But we can't simply remove the table, because if people use the
search-inside-table function and that was the result of the empty table they
loose the ability to change the search term.

We could only draw the table in the first place if there will be something in
it. But just finding out whether there will be something in the table can be
quite expensive, for instance with searches for values. So we don't want to do
this twice (once in the template to find out whether to draw the table in the
first place and once in the API call filling the table.) So we are currently
stuck with what we have until somebody invents a better way.