Add new relations section
There is a whole new relations section which gives information about different
relation types. There is a list of all relation types and there is a page for
each relation type which currently contains an empty overview tab and a roles
tab with a list of roles used with this relation type. There is currently no
link that leads to this section as it is not finished yet.