Hide unfinished reports.
Reports without sources are now "hidden", ie. they are not visible in the list
of reports. But if you know the URL you can access them. This allows them to be
tested before they are publicly released. Similarly APIs being tested can now
use the API version 0, which will make them invisible in the API documentation.