Fix a bug with nodes outside bbox
If there are nodes with locations outside the bounding box we are interested in
in the input osm file and we are generating the geodistribution for ways images,
tagstats could segfault. In this case the node location wasn't stored in the
m_storage, so when reading the it back, either you get an unintialized value
or a segfault. (This also depends on the type of storage used.)
Thanks to @lonvia for help debugging this.

The solution is to store "MAXINT" as a marker and make sure it doesn't end up
in the GeoDistribution.