Add code to create maps for tags.
For a long time we had the capability to create overview maps for keys, but
never for tags (ie key-value-combinations). This commit now adds code to
create maps for frequently used tags. We can't create maps for all tags,
because each map takes about 8k in RAM and there are a lot of tags.