Use sqlite db file instead of text *.lst files for selected data.
This changes the way some data is given from one taginfo update run
to the next. Instead of plain text files an sqlite database is used.
Part of the update is a new version of the upstream sqlite.hpp file.

After this commit the update has to be run twice to get all data
updated properly.