Add "Level0 Editor" button to key and tag pages.
This allows easy editing of tags, for instance when there are typos. If you
click on this link an Overpass API call will be made to find the object IDs for
the objects with this key/tag and then the Level0 editor will be opened in a
new window with those objects. To not overwhelm the Overpass API or the Level0
editor, the link will only be active if there are less then a configurable
number of objects with the given key/tags.