Use different overpass turbo call when numbers of objects is small.
When there aren't too many objects with a given key or tag, use the new wizard
interface of overpass turbo, which runs automatically in global context and
automatically zooms to the results. This is much nicer and easier.
If there are too many objects, we keep the old behaviour. Unfortunately this
is rather confusing. Maybe we can figure out something better...