Use pair of strings instead of concatenated string for hash map.
For the geodistribution hash we concatenated the key and value with an equal
sign in between and used that as a hash key. But this will lead to problems
when the key or value contain an equals sign. So instead we do the proper
thing now and use a std::pair<const char*, const char*>.