make sqlite3 full-text search tokenizer configurable.
OpenStreetMap is an international project. "icu" or other non-default
tokenizer may be more suitable for some locales. For example, "icu"
tokenizer is better than the default ("simple") for Chinese.

We don't want to force all to recompile sqlite3 in order to use "icu"
tokenizer, so make it configurable in taginfo-config.json.