Make JOSM button work in https mode.
The JOSM button now contacts JOSM on http://localhost:8111/ if the taginfo page
is using http and https://localhost:8112/ if it is using https. This way we
don't have a problem with mixed content.
Also the way JOSM is called now works without the iframes which were a hack
around the (then) missing CORS ability.