Some experimental code to find unicode properties in keys.
OSM allows nearly all unicode characters in keys and values. This is some
experimental code to find keys that contain unusual characters. Currently
this is not run or used in the normal setup, but I thought I'd throw it
out there in case somebody wants to play with it. I am not sure yet myself
where this is going.