Consistent access to in_wiki columns in database.
They are all checked with "to_i != 0" to turn them into bools.
Check will work in the future if/when we change the in_wiki column
to contain not just a boolean, but return the number of wiki pages.
This commit changes the output of the tags/list API call to return
a boolean instead of integer for the in_wiki attribute. The boolean
type was documented, so this now conforms to documentation.