Making the UI code a bit more modular.
The idea is that the admin can decide which parts of the UI to show.
This can be used to not show parts that don't make sense for a smaller
taginfo site or so. The current code only allows limited changes:
The /test, /taginfo, and /download sections can be switched on or off.
Note that if you switch off the /taginfo section, dangling links will
currently remain, so don't do that.