Use external content option for full text search index
Use the external content table option for the fts5 full text search
index. This way we need several GB less for the databases. To use
this we need to move the ftsearch table from the taginfo-search.db
into taginfo-db.db and don't need the taginfo-search.db at all
any more.

https://sqlite.org/fts5.html#external_content_and_contentless_tables