Reorganized Chinese language support
We support two versions, the traditional Chinese (zh-Hant, for instance
used in Taiwan) and the simplified Chinese (zh-Hans, for instance used
in mainland China). The ISO language codes zh-Hant and zh-Hans are used
in the OSM wiki for these languages (albeit in all lowercase). But
browsers use zh-TW and zh-CN (among some other codes) when requesting
the language. That's why we have those names in the list too and have
links from the i18n text files.