Large refactor: Get rid of flexigrid tables and jQuery
This removes the last remnants of jQuery. For this we had to reimplement
(in a much better way) the dynamic tables that were using the Flexigrid
library that hasn't been maintained in a very long time.

This also changes the CSS layout to flex and grid basically everywhere.
This means we don't have to use Javascript any more to resize HTML
elements but can let the browser do the right thing.

There is still a lot of code to clean up here. For instance the
"flexigrid" name survives in some places. And the Javascript is quite
horrible in many places...