Move project_unique_keys/tags to master (step 1)
The tables project_unique_keys and project_unique_tags are in the
projects database but they are changed by the master.sql script. This
prevents the projects database to be shared between taginfo instances.

This commit also adds those tables to the master database. It is step 1
of moving them over. In the next step we are changing the UI/API to use
the new tables and then we can remove the old tables later.