Use local copy of sinatra-r18n gem
The sinatra-r18n gem has not been updated in a long time. It works
okay with newer Sinatra versions but claims not to which prevents
us from updating Sinatra. Using the local copy solves this problem.

Also install puma web server instead of using webrick which is
kind of deprecated.